fix: fixed linting
