+++ lib/views/after_auth_screens/events/event_calendar.dart
@@ -3,21 +3,33 @@ import 'package:syncfusion_flutter_calendar/calendar.dart';
 import 'package:syncfusion_flutter_datepicker/datepicker.dart';
 import 'package:talawa/constants/routing_constants.dart';
 import 'package:talawa/locator.dart';
-import 'package:talawa/models/events/event_model.dart';
 import 'package:talawa/utils/app_localization.dart';
 import 'package:talawa/view_model/after_auth_view_models/event_view_models/event_calendar_view_model.dart';
+import 'package:talawa/views/after_auth_screens/events/event_calendar_helpers.dart';
+import 'package:talawa/views/after_auth_screens/events/event_card_widget.dart';
 import 'package:talawa/views/base_view.dart';
 import 'package:talawa/widgets/date_time_picker.dart';
 
+/// Display modes for the Explore Events screen.
+///
+/// `calendar` shows the month grid with a date-filtered list; `card` shows
+/// every loaded event grouped by day in a scrollable list.
+enum _EventViewMode {
+  /// Month grid with selected-day events listed below.
+  calendar,
+
+  /// Flat chronological list of all loaded events grouped by day.
+  card,
+}
+
 /// EventCalendar widget for displaying events in a month view calendar.
 class EventCalendar extends StatefulWidget {
+  /// Standard const constructor for the screen.
   const EventCalendar({super.key});
 
   @override
   State<EventCalendar> createState() => _EventCalendarState();
 }
-
-enum _EventViewMode { calendar, card }
 
 class _EventCalendarState extends State<EventCalendar> {
   DateTime? _selectedDate;
@@ -28,103 +40,123 @@ class _EventCalendarState extends State<EventCalendar> {
     return BaseView<EventCalendarViewModel>(
       onModelReady: (model) => model.initialize(),
       builder: (context, model, child) {
-        final appointments = _convertEventsToAppointments(model.eventList);
-        final List<Appointment> selectedDateEvents;
-        if (_selectedDate != null) {
-          final selStart = DateTime(
-            _selectedDate!.year,
-            _selectedDate!.month,
-            _selectedDate!.day,
-          );
-          final selEnd = selStart.add(const Duration(days: 1));
-          // Include any event whose range overlaps the selected day
-          // (covers genuine multi-day events while ignoring same-day all-day
-          // events whose end was clamped to 23:59 in the converter).
-          selectedDateEvents = appointments
-              .where(
-                (a) =>
-                    a.startTime.isBefore(selEnd) && a.endTime.isAfter(selStart),
-              )
-              .toList();
-        } else {
-          selectedDateEvents = const [];
-        }
+        final appointments = convertEventsToAppointments(model.eventList);
+        final selectedDateEvents = _selectedDateAppointments(appointments);
 
         return Scaffold(
-          appBar: AppBar(
-            backgroundColor: Colors.green,
-            elevation: 0.0,
-            centerTitle: true,
-            title: Text(
-              AppLocalizations.of(context)!.strictTranslate("Explore Events"),
-              style: Theme.of(context).textTheme.titleLarge!.copyWith(
-                    fontSize: 20,
-                    color: Colors.white,
-                  ),
-            ),
-            leading: IconButton(
-              icon: Icon(
-                Icons.menu,
-                color: Theme.of(context).iconTheme.color,
-              ),
-              onPressed: () => Scaffold.maybeOf(context)?.openDrawer(),
-            ),
-            actions: [
-              IconButton(
-                key: const Key('toggleEventViewMode'),
-                onPressed: () {
-                  setState(() {
-                    _viewMode = _viewMode == _EventViewMode.calendar
-                        ? _EventViewMode.card
-                        : _EventViewMode.calendar;
-                  });
-                },
-                icon: Icon(
-                  _viewMode == _EventViewMode.calendar
-                      ? Icons.view_agenda_outlined
-                      : Icons.calendar_month,
-                ),
-                tooltip: _viewMode == _EventViewMode.calendar
-                    ? 'Switch to card view'
-                    : 'Switch to calendar view',
-              ),
-              if (_viewMode == _EventViewMode.calendar)
-                IconButton(
-                  onPressed: () async {
-                    final pickedDate = await customDatePicker(
-                      initialDate: DateTime.now(),
-                    );
-                    model.selectionChanged(
-                      DateRangePickerSelectionChangedArgs(pickedDate),
-                    );
-                  },
-                  icon: const Icon(Icons.date_range),
-                  tooltip: 'Select Date',
-                ),
-              IconButton(
-                onPressed: () {
-                  navigationService.pushScreen(Routes.eventPageForm);
-                },
-                icon: const Icon(Icons.add),
-                tooltip: 'Add Event',
-              ),
-            ],
-          ),
+          appBar: _buildAppBar(context, model),
           body: _viewMode == _EventViewMode.calendar
-              ? _buildCalendarView(
-                  context,
-                  model,
-                  appointments,
-                  selectedDateEvents,
-                )
+              ? _buildCalendarView(model, appointments, selectedDateEvents)
               : _buildCardView(context, model, appointments),
         );
       },
     );
   }
 
+  /// Filters [appointments] to those overlapping the currently selected day.
+  ///
+  /// Returns an empty list when no day is selected. Range overlap covers
+  /// genuine multi-day events while ignoring same-day all-day events whose
+  /// end has been clamped to 23:59 in [convertEventsToAppointments].
+  ///
+  /// **params**:
+  /// * `appointments`: All loaded appointments.
+  ///
+  /// **returns**:
+  /// * `List<Appointment>`: Subset overlapping [_selectedDate].
+  List<Appointment> _selectedDateAppointments(List<Appointment> appointments) {
+    if (_selectedDate == null) return const [];
+    final selStart = DateTime(
+      _selectedDate!.year,
+      _selectedDate!.month,
+      _selectedDate!.day,
+    );
+    final selEnd = selStart.add(const Duration(days: 1));
+    return appointments
+        .where(
+          (a) => a.startTime.isBefore(selEnd) && a.endTime.isAfter(selStart),
+        )
+        .toList();
+  }
+
+  /// Builds the screen's [AppBar] including the view-mode toggle.
+  ///
+  /// **params**:
+  /// * `context`: Build context for theming and translations.
+  /// * `model`: Active [EventCalendarViewModel] for navigation callbacks.
+  ///
+  /// **returns**:
+  /// * `AppBar`: Configured app bar widget.
+  AppBar _buildAppBar(BuildContext context, EventCalendarViewModel model) {
+    return AppBar(
+      backgroundColor: Colors.green,
+      elevation: 0.0,
+      centerTitle: true,
+      title: Text(
+        AppLocalizations.of(context)!.strictTranslate("Explore Events"),
+        style: Theme.of(context).textTheme.titleLarge!.copyWith(
+              fontSize: 20,
+              color: Colors.white,
+            ),
+      ),
+      leading: IconButton(
+        icon: Icon(Icons.menu, color: Theme.of(context).iconTheme.color),
+        onPressed: () => Scaffold.maybeOf(context)?.openDrawer(),
+      ),
+      actions: [
+        IconButton(
+          key: const Key('toggleEventViewMode'),
+          onPressed: () {
+            setState(() {
+              _viewMode = _viewMode == _EventViewMode.calendar
+                  ? _EventViewMode.card
+                  : _EventViewMode.calendar;
+            });
+          },
+          icon: Icon(
+            _viewMode == _EventViewMode.calendar
+                ? Icons.view_agenda_outlined
+                : Icons.calendar_month,
+          ),
+          tooltip: _viewMode == _EventViewMode.calendar
+              ? 'Switch to card view'
+              : 'Switch to calendar view',
+        ),
+        if (_viewMode == _EventViewMode.calendar)
+          IconButton(
+            onPressed: () async {
+              final pickedDate =
+                  await customDatePicker(initialDate: DateTime.now());
+              model.selectionChanged(
+                DateRangePickerSelectionChangedArgs(pickedDate),
+              );
+            },
+            icon: const Icon(Icons.date_range),
+            tooltip: 'Select Date',
+          ),
+        IconButton(
+          onPressed: () => navigationService.pushScreen(Routes.eventPageForm),
+          icon: const Icon(Icons.add),
+          tooltip: 'Add Event',
+        ),
+      ],
+    );
+  }
+
+  /// Builds the month-calendar body for the Explore Events screen.
+  ///
+  /// Used when [_viewMode] is `calendar`; shows a month grid plus a list of
+  /// events for the currently selected day.
+  ///
+  /// **params**:
+  /// * `model`: Active [EventCalendarViewModel].
+  /// * `appointments`: All loaded events as SfCalendar appointments.
+  /// * `selectedDateEvents`: Subset of [appointments] overlapping the
+  ///   currently selected day.
+  ///
+  /// **returns**:
+  /// * `Widget`: The calendar-mode body.
   Widget _buildCalendarView(
-    BuildContext context,
     EventCalendarViewModel model,
     List<Appointment> appointments,
     List<Appointment> selectedDateEvents,
@@ -155,8 +187,10 @@ class _EventCalendarState extends State<EventCalendar> {
               Expanded(
                 child: ListView.builder(
                   itemCount: selectedDateEvents.length,
-                  itemBuilder: (context, index) =>
-                      _eventCard(context, model, selectedDateEvents[index]),
+                  itemBuilder: (context, index) => EventCardWidget(
+                    model: model,
+                    event: selectedDateEvents[index],
+                  ),
                 ),
               ),
             ],
@@ -166,6 +200,18 @@ class _EventCalendarState extends State<EventCalendar> {
     );
   }
 
+  /// Builds the chronological card list of events.
+  ///
+  /// Used when [_viewMode] is `card`; renders every loaded event grouped by
+  /// day in a scrollable list.
+  ///
+  /// **params**:
+  /// * `context`: Build context for theming.
+  /// * `model`: Active [EventCalendarViewModel].
+  /// * `appointments`: All loaded events as SfCalendar appointments.
+  ///
+  /// **returns**:
+  /// * `Widget`: The card-mode body.
   Widget _buildCardView(
     BuildContext context,
     EventCalendarViewModel model,
@@ -220,10 +266,7 @@ class _EventCalendarState extends State<EventCalendar> {
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
             Padding(
-              padding: const EdgeInsets.symmetric(
-                vertical: 8,
-                horizontal: 4,
-              ),
+              padding: const EdgeInsets.symmetric(vertical: 8, horizontal: 4),
               child: Text(
                 header,
                 style: Theme.of(context).textTheme.titleSmall!.copyWith(
@@ -232,217 +275,10 @@ class _EventCalendarState extends State<EventCalendar> {
                     ),
               ),
             ),
-            ...dayEvents.map((a) => _eventCard(context, model, a)),
+            ...dayEvents.map((a) => EventCardWidget(model: model, event: a)),
           ],
         );
       },
     );
-  }
-
-  Widget _eventCard(
-    BuildContext context,
-    EventCalendarViewModel model,
-    Appointment event,
-  ) {
-    final String startTime =
-        '${event.startTime.hour.toString().padLeft(2, '0')}:${event.startTime.minute.toString().padLeft(2, '0')}';
-    final String endTime =
-        '${event.endTime.hour.toString().padLeft(2, '0')}:${event.endTime.minute.toString().padLeft(2, '0')}';
-
-    return GestureDetector(
-      onTap: () {
-        final originalEvent =
-            model.eventList.firstWhere((e) => e.id == event.id);
-        navigationService.pushScreen(
-          "/eventInfo",
-          arguments: originalEvent,
-        );
-      },
-      child: Card(
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(12),
-        ),
-        margin: const EdgeInsets.symmetric(vertical: 6),
-        elevation: 2,
-        child: Padding(
-          padding: const EdgeInsets.symmetric(
-            vertical: 12,
-            horizontal: 16,
-          ),
-          child: Row(
-            children: [
-              Container(
-                width: 5,
-                height: 50,
-                decoration: BoxDecoration(
-                  color: event.color,
-                  borderRadius: BorderRadius.circular(4),
-                ),
-              ),
-              const SizedBox(width: 12),
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      event.subject,
-                      style: Theme.of(context).textTheme.titleMedium!.copyWith(
-                            fontWeight: FontWeight.w600,
-                            fontSize: 16,
-                          ),
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                    const SizedBox(height: 6),
-                    Row(
-                      children: [
-                        Icon(
-                          Icons.access_time,
-                          size: 16,
-                          color: Colors.grey[700],
-                        ),
-                        const SizedBox(width: 4),
-                        Text(
-                          '$startTime - $endTime',
-                          style: Theme.of(context)
-                              .textTheme
-                              .bodyMedium!
-                              .copyWith(color: Colors.grey[700]),
-                        ),
-                      ],
-                    ),
-                    if ((event.location ?? '').isNotEmpty) ...[
-                      const SizedBox(height: 4),
-                      Row(
-                        children: [
-                          Icon(
-                            Icons.location_on_outlined,
-                            size: 16,
-                            color: Colors.grey[700],
-                          ),
-                          const SizedBox(width: 4),
-                          Expanded(
-                            child: Text(
-                              event.location!,
-                              style: Theme.of(context)
-                                  .textTheme
-                                  .bodyMedium!
-                                  .copyWith(color: Colors.grey[700]),
-                              overflow: TextOverflow.ellipsis,
-                            ),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ],
-                ),
-              ),
-              const Icon(
-                Icons.chevron_right,
-                color: Colors.grey,
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-/// Converts Event objects to Appointment objects for the calendar.
-///
-/// Includes validation and fallbacks for:
-/// - Null or invalid start/end dates
-/// - End time before start time
-/// - Zero duration events
-/// - Invalid date ranges
-///
-/// **params**:
-/// * `eventsList`: List of Event objects to convert
-///
-/// **returns**:
-/// * `List<Appointment>`: List of valid Appointment objects for the calendar
-List<Appointment> _convertEventsToAppointments(List<Event> eventsList) {
-  const colors = [
-    Colors.green,
-    Colors.blue,
-    Colors.red,
-    Colors.orange,
-    Colors.purple,
-    Colors.pink,
-  ];
-  final List<Appointment> appointments = [];
-
-  for (final event in eventsList) {
-    try {
-      // Skip events without required date fields
-      if (event.startAt == null || event.endAt == null) {
-        continue;
-      }
-
-      final index = eventsList.indexOf(event);
-
-      // Create DateTime objects with validation
-      final DateTime startDate = DateTime(
-        event.startAt!.year,
-        event.startAt!.month,
-        event.startAt!.day,
-        event.startAt!.hour,
-        event.startAt!.minute,
-        event.startAt!.second,
-      );
-
-      DateTime endDate = DateTime(
-        event.endAt!.year,
-        event.endAt!.month,
-        event.endAt!.day,
-        event.endAt!.hour,
-        event.endAt!.minute,
-        event.endAt!.second,
-      );
-
-      // The API rejects same-day all-day events, so we bump endDate by +1 on
-      // create (see CreateEventViewModel). Clamp the appointment end back to
-      // the start day for display so the calendar shows the dot on a single
-      // cell and the date-filtered list picks it up correctly.
-      final isAllDay = event.allDay ?? false;
-      if (isAllDay &&
-          endDate.difference(startDate).inDays == 1 &&
-          endDate.hour == 0 &&
-          endDate.minute == 0) {
-        endDate = DateTime(
-          startDate.year,
-          startDate.month,
-          startDate.day,
-          23,
-          59,
-          59,
-        );
-      }
-
-      final appointment = Appointment(
-        startTime: startDate,
-        endTime: endDate,
-        subject: event.name?.isNotEmpty == true ? event.name! : 'Unnamed Event',
-        color: colors[index % colors.length],
-        location: event.location ?? '',
-        id: event.id,
-        isAllDay: isAllDay,
-      );
-
-      appointments.add(appointment);
-    } catch (e) {
-      // Catch any unexpected errors during conversion
-      debugPrint('Error converting event ${event.id} to appointment: $e');
-      continue;
-    }
-  }
-
-  return appointments;
-}
-
-/// Simple data source for SfCalendar.
-class EventDataSource extends CalendarDataSource {
-  EventDataSource(List<Appointment> appointments) {
-    this.appointments = appointments;
   }
 }
